Refactor GCP and Static to be far more generic
I had all these ideas about how I'd use an `IMedia` interface to describe all
different kinds of media files, but really... the only thing that varies is the
MIME type, and that's much better passed as configuration to a record. So!

This strips out IMedia and some of the foofaraw around it, replacing it with a
`Media` record. On the `Static` side, we declare a type called `Upload` and go
from there. Configuration can be passed in by the calling function, in a
pleasing way.