Rewrite in Giraffe
This is a direct re-implementation of Cookbook in Giraffe. Visually, it shows
almost no difference at all -- which was the goal. It is .NET5.0 and very
current -- and, honestly, cleaner and slimmer and clearer than anything I did in
WS.

From here, a great deal of further refinement is possible. There's a great deal
of async buried in the Dropbox package especially, and refactoring that more
sensibly, cleaning up the HTTP handlers, will make the whole works a deal easier
to reason about.