Mostly get static asset syncing working, mostly
This isn't... perfect, but most of the parts are there. Just one weird catch: it
doesn't appear to be reading my local credentials correctly. So, either this is
gonna work amazingly super perfect in prod, or it's gonna error all night until
I can fix it. Might as well test it!