Refactor the GCP client to be a class over a record, add to DI
This changes the GCP client to be a class, which has a normal constructor, which
can be added to DI. I feel pretty weird about "add it to DI" as a premise, but
here we are. It's the "normal" way to share this kind of data between the system
and business logic code.