Split apart config files, fix DI, inject file client
OK, so this is once again progress. It was... opaque, to be honest, but it turns
out, you *have* to register your config class with DI. Mine was only working
because I provided sensible default values (a mistake in retrospect). This gets
the config class hooked to DI. Further, conditionally wire up a file client in dev.