Re-configure most stuff
This scoots around a bunch of things into smaller units in better-labeled/named
places. I broke out a new Web module, which I'm fairly pleased with. There's
likely just a little more to do before the factoring in here is really correct,
but I'm in to it so far. Got the tests tidied and passing, too.