Heap Cache, step one: start untangling IStorageClient
This starts getting memory caching wired up. Adding a memory cache to ASP.NET
turned out to be trivially easy. Dealing with my past code choices is taking
more work to unwind.

I made the fairly natural decision to stash my `IStorageClient` interface and
impl in my GCP module... which is not helping. That's just not gonna help. Nope.
No help. So, we'll move that interface into Cookbook Proper and figure out how
we're gonna Deal With Things from there.

This work is... far from done. But this is an excellent step.