Implement tag manifest loading, protect XML parsing
We want to parse tag manifests from org-blorg, which will be authoritative lists
of all the tags used in a project. This is a lovely idea, and also, it turned up
some _stuff_.

First, very practically, we add Thoth for JSON parsing. I could do this a
zillion ways, but Thoth is nice. I don't even need a wrapper type for Tags, so
far. Tag manifest parsing can then be wired up and percolated through, which requires
Result handling. Not bad.

But: while doing this, I learned that FSharp.Data's XML parser will *detonate*
if you, say, accidentally feed it JSON -- and the errors are horrible and very,
very opaque. So, I wrapped that in a Result as well, which requires a bit more
percolating through the stack. Should be... I dunno, at least a lot less
dramatic now.

This is all set, once tags are fixed on my blog!