Improve CSS for a variety of elements
This is progress, though definitely not complete, alas.

This improves CSS for `dl` elements considerably, which was not hard because
they were previously totally unstyled. I've made `hr` elements inside post
bodies look considerably nicer.

I tried to fix the footnotes section, but maddeningly, I cannot get them to
behave. Partly, this is because of the _ridiculous_ way Org exports the
footnotes section, wrapping the footnote number in a `sup` and the body in a
frigging `div`. Ugh. I'll need to update org-blorg to put this stuff in a `span`
before I can get the look I actually want.